+++ src/analysis/import/import_neighborhood.sh
@@ -66,13 +66,9 @@ then
         usage
     else
         NB_BOUNDARY_FILE="${1}"
-        if [[ -n "${3}" ]]
-          then
-            NB_STATE="${2}"
-            NB_STATE_FIPS="${3}"
-        else
-            NB_COUNTRY="${2}"
-        fi
+        NB_COUNTRY="${2}"
+        NB_STATE="${3}"
+        NB_STATE_FIPS="${4}"
 
         NB_TEMPDIR="${NB_TEMPDIR:-$(mktemp -d)}/import_neighborhood"
         mkdir -p "${NB_TEMPDIR}"
@@ -83,8 +79,7 @@ then
         update_status "IMPORTING" "Importing boundary shapefile"
         import_and_transform_shapefile "${NB_BOUNDARY_FILE}" neighborhood_boundary "${NB_INPUT_SRID}"
 
-        if [ -n "$NB_STATE_FIPS" ]
-          then
+        if [ "${PFB_COUNTRY}" == "USA" ]; then
             update_status "IMPORTING" "Downloading water blocks"
             # Create water blocks table
             psql -h "${NB_POSTGRESQL_HOST}" -U "${NB_POSTGRESQL_USER}" -d "${NB_POSTGRESQL_DB}" \
@@ -120,8 +115,9 @@ then
 
         # Get blocks for the place requested
         update_status "IMPORTING" "Downloading census blocks"
-        if [[ -n "${NB_STATE_FIPS}" ]]; then
+        if [ "${PFB_COUNTRY}" == "USA" ]; then
             NB_BLOCK_FILENAME="tabblock2010_${NB_STATE_FIPS}_pophu"
+            PFB_POP_URL="http://www2.census.gov/geo/tiger/TIGER2010BLKPOPHU/${NB_BLOCK_FILENAME}.zip"
         else
             NB_BLOCK_FILENAME="population"
         fi
@@ -130,19 +126,19 @@ then
         if [ -f "/data/${NB_BLOCK_FILENAME}.zip" ]; then
             echo "Using local census blocks file"
             BLOCK_DOWNLOAD="/data/${NB_BLOCK_FILENAME}.zip"
-        elif [ -n "${PFB_POP_URL}" ]; then
-            echo "Using blocks file from PFB_POP_URL: ${PFB_POP_URL}"
-            BLOCK_DOWNLOAD="${NB_TEMPDIR}/${NB_BLOCK_FILENAME}.zip"
-            wget -nv -O "${BLOCK_DOWNLOAD}" "${PFB_POP_URL}"
         elif [ "${AWS_STORAGE_BUCKET_NAME}" ] && aws s3 ls "${S3_PATH}"; then
             echo "Using census blocks file from S3"
             BLOCK_DOWNLOAD="${NB_TEMPDIR}/${NB_BLOCK_FILENAME}.zip"
             aws s3 cp "${S3_PATH}" "${BLOCK_DOWNLOAD}"
         else
-            echo "Using census blocks file from official census site"
             BLOCK_DOWNLOAD="${NB_TEMPDIR}/${NB_BLOCK_FILENAME}.zip"
-            PFB_POP_URL="http://www2.census.gov/geo/tiger/TIGER2010BLKPOPHU/${NB_BLOCK_FILENAME}.zip"
-            wget -nv -O "${BLOCK_DOWNLOAD}" "${PFB_POP_URL}"
+            if [ "${PFB_COUNTRY}" == "USA" ]; then
+                echo "Using census blocks file from official census site"
+            else
+                echo "Using blocks file from PFB_POP_URL: ${PFB_POP_URL}"
+            fi
+            wget -nv -O "${BLOCK_DOWNLOAD}" "${PFB_POP_URL}"      
+
             if [ "${AWS_STORAGE_BUCKET_NAME}" ]; then
                 echo "Uploading census blocks file to S3 cache"
                 aws s3 cp "${BLOCK_DOWNLOAD}" "${S3_PATH}"
@@ -150,11 +146,13 @@ then
         fi
         unzip "${BLOCK_DOWNLOAD}" -d "${NB_TEMPDIR}"
 
-        # Rename unzipped files so they can be found easily
-        cd $NB_TEMPDIR
-        rm population.zip
-        for x in *; do mv "$x" "${NB_BLOCK_FILENAME}.${x##*.}"; done
-        cd - 
+        if [ "${PFB_COUNTRY}" != "USA" ]; then
+            # Rename unzipped files if not from census so they can be found easily
+            cd $NB_TEMPDIR
+            rm "${NB_BLOCK_FILENAME}.zip"
+            for x in *; do mv "$x" "${NB_BLOCK_FILENAME}.${x##*.}"; done
+            cd - 
+        fi
 
         # Import block shapefile
         update_status "IMPORTING" "Loading census blocks"
@@ -169,8 +167,7 @@ then
                 ${NB_BOUNDARY_BUFFER});"
         echo "DONE: Finished removing blocks outside buffer"
         
-        if [ -n "$NB_STATE_FIPS" ]
-          then
+        if [ "${PFB_COUNTRY}" == "USA" ]; then
             # Discard blocks that are all water / no land area
             update_status "IMPORTING" "Removing water blocks"
             echo "START: Removing blocks that are 100% water from analysis"
