--- src/analysis/scripts/continents_to_country.py
@@ -1,256 +1,0 @@
-#!/usr/bin/env python3
-
-country_to_continent = {
-"ABW": "",
-"AFG": "Asia",
-"AGO": "Africa",
-"AIA": "",
-"ALA": "",
-"ALB": "Europe",
-"AND": "Europe",
-"ARE": "Asia",
-"ARG": "South America",
-"ARM": "Europe",
-"ASM": "",
-"ATA": "",
-"ATF": "",
-"ATG": "North America",
-"AUS": "Australia Oceania",
-"AUT": "Europe",
-"AZE": "Europe",
-"BDI": "Africa",
-"BEL": "Europe",
-"BEN": "Africa",
-"BES": "",
-"BFA": "",
-"BGD": "Asia",
-"BGR": "Europe",
-"BHR": "Asia",
-"BHS": "Central America",
-"BIH": "Europe",
-"BLM": "",
-"BLR": "Europe",
-"BLZ": "Central America",
-"BMU": "",
-"BOL": "",
-"BRA": "South America",
-"BRB": "North America",
-"BRN": "Asia",
-"BTN": "Asia",
-"BVT": "",
-"BWA": "Africa",
-"CAF": "Africa",
-"CAN": "North America",
-"CCK": "",
-"CHE": "Europe",
-"CHL": "South America",
-"CHN": "Asia",
-"CIV": "",
-"CMR": "Africa",
-"COD": "",
-"COG": "Africa",
-"COK": "",
-"COL": "South America",
-"COM": "Africa",
-"CPV": "",
-"CRI": "Central America",
-"CUB": "Central America",
-"CUW": "",
-"CXR": "",
-"CYM": "",
-"CYP": "Europe",
-"CZE": "",
-"DEU": "Europe",
-"DJI": "Africa",
-"DMA": "North America",
-"DNK": "Europe",
-"DOM": "Central America",
-"DZA": "Africa",
-"ECU": "South America",
-"EGY": "Africa",
-"ERI": "Africa",
-"ESH": "",
-"ESP": "Europe",
-"EST": "Europe",
-"ETH": "Africa",
-"FIN": "Europe",
-"FJI": "Australia Oceania",
-"FLK": "",
-"FRA": "Europe",
-"FRO": "",
-"FSM": "",
-"GAB": "Africa",
-"GBR": "Europe",
-"GEO": "Europe",
-"GGY": "",
-"GHA": "Africa",
-"GIB": "",
-"GIN": "Africa",
-"GLP": "",
-"GMB": "Africa",
-"GNB": "Africa",
-"GNQ": "Africa",
-"GRC": "Europe",
-"GRD": "North America",
-"GRL": "North America",
-"GTM": "Central America",
-"GUF": "",
-"GUM": "",
-"GUY": "South America",
-"HKG": "",
-"HMD": "",
-"HND": "Central America",
-"HRV": "Europe",
-"HTI": "Central America",
-"HUN": "Europe",
-"IDN": "Asia",
-"IMN": "",
-"IND": "Asia",
-"IOT": "",
-"IRL": "Europe",
-"IRN": "",
-"IRQ": "Asia",
-"ISL": "Europe",
-"ISR": "Asia",
-"ITA": "Europe",
-"JAM": "Central America",
-"JEY": "",
-"JOR": "Asia",
-"JPN": "Asia",
-"KAZ": "Asia",
-"KEN": "Africa",
-"KGZ": "Asia",
-"KHM": "Asia",
-"KIR": "Australia Oceania",
-"KNA": "North America",
-"KOR": "",
-"KWT": "Asia",
-"LAO": "",
-"LBN": "Asia",
-"LBR": "Africa",
-"LBY": "Africa",
-"LCA": "North America",
-"LIE": "Europe",
-"LKA": "Asia",
-"LSO": "Africa",
-"LTU": "Europe",
-"LUX": "Europe",
-"LVA": "Europe",
-"MAC": "",
-"MAF": "",
-"MAR": "Africa",
-"MCO": "Europe",
-"MDA": "",
-"MDG": "Africa",
-"MDV": "Asia",
-"MEX": "North America",
-"MHL": "Australia Oceania",
-"MKD": "",
-"MLI": "Africa",
-"MLT": "Europe",
-"MMR": "",
-"MNE": "Europe",
-"MNG": "Asia",
-"MNP": "",
-"MOZ": "Africa",
-"MRT": "Africa",
-"MSR": "",
-"MTQ": "",
-"MUS": "Africa",
-"MWI": "Africa",
-"MYS": "Asia",
-"MYT": "",
-"NAM": "Africa",
-"NCL": "",
-"NER": "Africa",
-"NFK": "",
-"NGA": "Africa",
-"NIC": "Central America",
-"NIU": "",
-"NLD": "Europe",
-"NOR": "Europe",
-"NPL": "Asia",
-"NRU": "Australia Oceania",
-"NZL": "Australia Oceania",
-"OMN": "Asia",
-"PAK": "Asia",
-"PAN": "Central America",
-"PCN": "",
-"PER": "South America",
-"PHL": "Asia",
-"PLW": "Australia Oceania",
-"PNG": "Australia Oceania",
-"POL": "Europe",
-"PRI": "",
-"PRK": "",
-"PRT": "Europe",
-"PRY": "South America",
-"PSE": "",
-"PYF": "",
-"QAT": "Asia",
-"REU": "",
-"ROU": "Europe",
-"RUS": "Asia",
-"RWA": "Africa",
-"SAU": "Asia",
-"SDN": "Africa",
-"SEN": "Africa",
-"SGP": "Asia",
-"SGS": "",
-"SHN": "",
-"SJM": "",
-"SLB": "Australia Oceania",
-"SLE": "Africa",
-"SLV": "Central America",
-"SMR": "Europe",
-"SOM": "Africa",
-"SPM": "",
-"SRB": "Europe",
-"SSD": "Africa",
-"STP": "Africa",
-"SUR": "South America",
-"SVK": "Europe",
-"SVN": "Europe",
-"SWE": "Europe",
-"SWZ": "",
-"SXM": "",
-"SYC": "Africa",
-"SYR": "",
-"TCA": "",
-"TCD": "Africa",
-"TGO": "Africa",
-"THA": "Asia",
-"TJK": "Asia",
-"TKL": "",
-"TKM": "Asia",
-"TLS": "",
-"TON": "Australia Oceania",
-"TTO": "North America",
-"TUN": "Africa",
-"TUR": "Asia",
-"TUV": "Australia Oceania",
-"TWN": "",
-"TZA": "",
-"UGA": "Africa",
-"UKR": "Europe",
-"UMI": "",
-"URY": "South America",
-"USA": "North America",
-"UZB": "Asia",
-"VAT": "",
-"VCT": "North America",
-"VEN": "South America",
-"VGB": "",
-"VIR": "",
-"VNM": "",
-"VUT": "Australia Oceania",
-"WLF": "",
-"WSM": "Australia Oceania",
-"YEM": "Asia",
-"ZAF": "Africa",
-"ZMB": "Africa",
-"ZWE": "Africa"
-}
-
-def get_continent(country):
-    return globals()['country_to_continent'][country]
