+++ src/analysis/scripts/country_to_continent.py
@@ -1,0 +1,179 @@
+# Continents are used to derive Geofabrik extract URLs
+def get_continent(country):
+    return {
+        "AFG": "Asia",
+        "AGO": "Africa",
+        "ALB": "Europe",
+        "AND": "Europe",
+        "ARE": "Asia",
+        "ARG": "South America",
+        "ARM": "Europe",
+        "ATG": "North America",
+        "AUS": "Australia Oceania",
+        "AUT": "Europe",
+        "AZE": "Europe",
+        "BDI": "Africa",
+        "BEL": "Europe",
+        "BEN": "Africa",
+        "BGD": "Asia",
+        "BGR": "Europe",
+        "BHR": "Asia",
+        "BHS": "Central America",
+        "BIH": "Europe",
+        "BLR": "Europe",
+        "BLZ": "Central America",
+        "BRA": "South America",
+        "BRB": "North America",
+        "BRN": "Asia",
+        "BTN": "Asia",
+        "BWA": "Africa",
+        "CAF": "Africa",
+        "CAN": "North America",
+        "CHE": "Europe",
+        "CHL": "South America",
+        "CHN": "Asia",
+        "CMR": "Africa",
+        "COG": "Africa",
+        "COL": "South America",
+        "COM": "Africa",
+        "CRI": "Central America",
+        "CUB": "Central America",
+        "CYP": "Europe",
+        "DEU": "Europe",
+        "DJI": "Africa",
+        "DMA": "North America",
+        "DNK": "Europe",
+        "DOM": "Central America",
+        "DZA": "Africa",
+        "ECU": "South America",
+        "EGY": "Africa",
+        "ERI": "Africa",
+        "ESP": "Europe",
+        "EST": "Europe",
+        "ETH": "Africa",
+        "FIN": "Europe",
+        "FJI": "Australia Oceania",
+        "FRA": "Europe",
+        "GAB": "Africa",
+        "GBR": "Europe",
+        "GEO": "Europe",
+        "GHA": "Africa",
+        "GIN": "Africa",
+        "GMB": "Africa",
+        "GNB": "Africa",
+        "GNQ": "Africa",
+        "GRC": "Europe",
+        "GRD": "North America",
+        "GRL": "North America",
+        "GTM": "Central America",
+        "GUY": "South America",
+        "HND": "Central America",
+        "HRV": "Europe",
+        "HTI": "Central America",
+        "HUN": "Europe",
+        "IDN": "Asia",
+        "IND": "Asia",
+        "IRL": "Europe",
+        "IRQ": "Asia",
+        "ISL": "Europe",
+        "ISR": "Asia",
+        "ITA": "Europe",
+        "JAM": "Central America",
+        "JOR": "Asia",
+        "JPN": "Asia",
+        "KAZ": "Asia",
+        "KEN": "Africa",
+        "KGZ": "Asia",
+        "KHM": "Asia",
+        "KIR": "Australia Oceania",
+        "KNA": "North America",
+        "KWT": "Asia",
+        "LBN": "Asia",
+        "LBR": "Africa",
+        "LBY": "Africa",
+        "LCA": "North America",
+        "LIE": "Europe",
+        "LKA": "Asia",
+        "LSO": "Africa",
+        "LTU": "Europe",
+        "LUX": "Europe",
+        "LVA": "Europe",
+        "MAR": "Africa",
+        "MCO": "Europe",
+        "MDG": "Africa",
+        "MDV": "Asia",
+        "MEX": "North America",
+        "MHL": "Australia Oceania",
+        "MLI": "Africa",
+        "MLT": "Europe",
+        "MNE": "Europe",
+        "MNG": "Asia",
+        "MOZ": "Africa",
+        "MRT": "Africa",
+        "MUS": "Africa",
+        "MWI": "Africa",
+        "MYS": "Asia",
+        "NAM": "Africa",
+        "NER": "Africa",
+        "NGA": "Africa",
+        "NIC": "Central America",
+        "NLD": "Europe",
+        "NOR": "Europe",
+        "NPL": "Asia",
+        "NRU": "Australia Oceania",
+        "NZL": "Australia Oceania",
+        "OMN": "Asia",
+        "PAK": "Asia",
+        "PAN": "Central America",
+        "PER": "South America",
+        "PHL": "Asia",
+        "PLW": "Australia Oceania",
+        "PNG": "Australia Oceania",
+        "POL": "Europe",
+        "PRT": "Europe",
+        "PRY": "South America",
+        "QAT": "Asia",
+        "ROU": "Europe",
+        "RUS": "Asia",
+        "RWA": "Africa",
+        "SAU": "Asia",
+        "SDN": "Africa",
+        "SEN": "Africa",
+        "SGP": "Asia",
+        "SLB": "Australia Oceania",
+        "SLE": "Africa",
+        "SLV": "Central America",
+        "SMR": "Europe",
+        "SOM": "Africa",
+        "SRB": "Europe",
+        "SSD": "Africa",
+        "STP": "Africa",
+        "SUR": "South America",
+        "SVK": "Europe",
+        "SVN": "Europe",
+        "SWE": "Europe",
+        "SYC": "Africa",
+        "TCD": "Africa",
+        "TGO": "Africa",
+        "THA": "Asia",
+        "TJK": "Asia",
+        "TKM": "Asia",
+        "TON": "Australia Oceania",
+        "TTO": "North America",
+        "TUN": "Africa",
+        "TUR": "Asia",
+        "TUV": "Australia Oceania",
+        "UGA": "Africa",
+        "UKR": "Europe",
+        "URY": "South America",
+        "USA": "North America",
+        "UZB": "Asia",
+        "VCT": "North America",
+        "VEN": "South America",
+        "VUT": "Australia Oceania",
+        "WSM": "Australia Oceania",
+        "YEM": "Asia",
+        "ZAF": "Africa",
+        "ZMB": "Africa",
+        "ZWE": "Africa"
+    }.get(country, "")
