+++ src/analysis/scripts/import.sh
@@ -27,13 +27,13 @@ then
     else
         PFB_SHPFILE="${1}"
         PFB_OSM_FILE="${2}"
-        PFB_COUNTRY_OR_STATE="${3}"
-        PFB_STATE_FIPS="${4}"
+        PFB_COUNTRY="${3}"
+        PFB_STATE="${4}"
+        PFB_STATE_FIPS="${5}"
 
-        ../import/import_neighborhood.sh $PFB_SHPFILE $PFB_COUNTRY_OR_STATE $PFB_STATE_FIPS
-        if [ "$RUN_IMPORT_JOBS" = "1" ]
-        then
-            ../import/import_jobs.sh $PFB_COUNTRY_OR_STATE
+        ../import/import_neighborhood.sh $PFB_SHPFILE $PFB_COUNTRY $PFB_STATE $PFB_STATE_FIPS
+        if [ "$RUN_IMPORT_JOBS" == "1" ]; then
+            ../import/import_jobs.sh $PFB_COUNTRY $PFB_STATE
         else
             echo "Skipping Importing Jobs"
         fi
