+++ src/django/pfb_analysis/models.py
@@ -397,9 +397,9 @@ class AnalysisBatchManager(models.Manager):
                 # Handle NULL values in the city_fips property, which aren't allowed in the model
                 if city_fips is None:
                     city_fips = ''
-                osm_extract_url = feature['properties'].get('osm_url', None)
-                population_url = feature['properties'].get('population_url', None)
-                jobs_url = feature['properties'].get('jobs_url', None)
+                osm_extract_url = '' if feature['properties'].get('osm_url', '') is None else feature['properties'].get('osm_url', '')
+                population_url = '' if feature['properties'].get('population_url', '') is None else feature['properties'].get('population_url', '')
+                jobs_url = '' if feature['properties'].get('jobs_url', '') is None else feature['properties'].get('jobs_url', '')
                 skip_import_jobs = feature['properties'].get('skip_import_jobs', False)
                 label = city
                 name = Neighborhood.name_for_label(label)
@@ -755,14 +755,12 @@ class AnalysisJob(PFBModel):
             'PFB_JOB_ID': str(self.uuid),
             'AWS_STORAGE_BUCKET_NAME': settings.AWS_STORAGE_BUCKET_NAME,
             'PFB_S3_RESULTS_PATH': self.s3_results_path,
-            'PFB_POP_URL': self.population_url,
-            'PFB_JOB_URL': self.jobs_url
+            'PFB_JOB_URL': self.jobs_url if self.jobs_url is not None else '',
+            'PFB_OSM_FILE_URL': self.osm_extract_url if self.osm_extract_url is not None else '',
+            'PFB_POP_URL': self.population_url if self.population_url is not None else '',
+            'RUN_IMPORT_JOBS': 0 if self.skip_import_jobs else 1
         })
 
-        if self.osm_extract_url:
-            environment['PFB_OSM_FILE_URL'] = self.osm_extract_url
-        environment['RUN_IMPORT_JOBS'] = 0 if self.skip_import_jobs else 1
-   
         # Workaround for not being able to run development jobs on the actual batch cluster:
         # bail out with a helpful message
         if settings.DJANGO_ENV == 'development':
